backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
openai==1.58.1
weasyprint==63.1
python-dotenv==1.0.1
jinja2==3.1.5
pyproj==3.7.1
lxml==5.3.0
pydantic==2.10.4
pydantic-settings==2.7.1
